Wip merge serialized results list classes to one class, move lbp3 category endpoints to own class
